refactor: gridId 타입 수정 (long -> int)

+++ backend/application-modules/batch-server/src/main/java/com/softeer/batch/forecast/mountain/reader/MountainIdentifierReader.java
@@ -27,7 +27,7 @@ public class MountainIdentifierReader extends JdbcPagingItemReader<MountainIdent
         super.setRowMapper((rs, rowNum) -> new MountainIdentifier(
                 rs.getLong("id"),
                 rs.getInt("code"),
-                rs.getLong("gridId")
+                rs.getInt("gridId")
         ));
 
         try {
